+++ src/commands/build.rs
@@ -60,7 +60,7 @@ impl Subcommand<Context> for Build {
         cli.help(Help::with(build::HELP))?;
         Ok(Build {
             // Flags
-            list: cli.check(Arg::flag("list"))?,
+            list: cli.check(Arg::flag("list").switch('l'))?,
             verbose: cli.check(Arg::flag("verbose"))?,
             force: cli.check(Arg::flag("force"))?,
             all: cli.check(Arg::flag("all"))?,