+++ src/commands/helps/publish.rs
@@ -24,7 +24,7 @@ Usage:
 Options:
     --ready, -y           run the operation to completion
     --no-install          do not install the ip for future use
-    --list                view available channels and exit
+    --list, -l            view available channels and exit
     --all-public          publish with all source files being public
 
 Use 'orbit help publish' to read more about the command."#;